requirements.txt
DateTime==5.5
folium==0.16.0
haversine==2.8.1
matplotlib==3.9.0
matplotlib-inline==0.1.7
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
streamlit==1.35.0
streamlit_folium==0.20.0